third_party/blink/web_tests/platform/win/external/wpt/html/rendering/widgets/button-layout/grid-expected.txt

This is a testharness.js-based test.
FAIL inline-grid assert_equals: clientWidth expected 26 but got 19
FAIL grid assert_equals: clientWidth expected 26 but got 19
Harness: the test ran to completion.
